Move layer rendering API to item_rendering
The LayerRenderer trait and the render_layer() free function now live
next to ItemRenderer rather than on Layer. Layer stays item-only.
